requirements.txt
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==7.1.2
comm==0.2.1
contourpy==1.1.1
coverage==7.4.0
cryptography==41.0.7
cycler==0.12.1
Cython==3.0.8
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docutils==0.20.1
exceptiongroup==1.2.0
fastjsonschema==2.19.1
fonttools==4.25.0
fqdn==1.5.1
funcsigs==1.0.2
fuzzy-c-means==1.6.3
idna==3.6
implicit==0.7.2
importlib-metadata==7.0.1
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.28.0
ipython==7.34.0
ipywidgets==8.1.1
isoduration==20.11.0
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.1.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keyring==24.3.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
munkres==1.1.4
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
nh3==0.2.15
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.4
overrides==7.4.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.2.0
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
plotly==4.14.3
pluggy==1.3.0
ply==3.11
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.1
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.13
Pygments==2.17.2
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pytest==7.4.4
pytest-cov==2.12.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
readme-renderer==42.0
referencing==0.32.1
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.16.2
scikit-learn==1.3.2
scikit-surprise==1.1.3
scipy==1.9.0
seaborn==0.11.2
SecretStorage==3.3.3
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tabulate==0.8.10
terminado==0.18.0
threadpoolctl==3.5.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
twine==4.0.2
typer==0.3.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
zipp==3.19.1
